requirements.txt
setuptools~=75.8.0
torch~=2.6.0
matplotlib~=3.10.0
torchvision~=0.21.0
opencv-python~=4.11.0.86
numpy~=2.2.2
pillow~=11.1.0